+++ services/workflows-service/scripts/generate-end-user.ts
@@ -102,10 +102,11 @@ export const generateBusiness = ({
     workflowDefinitionId: string;
     workflowDefinitionVersion: number;
     context: Prisma.InputJsonValue;
+    state: string;
   };
   projectId: string;
 }): Prisma.BusinessCreateInput => {
-  const { runtimeId, workflowDefinitionId, workflowDefinitionVersion, context } = workflow;
+  const { runtimeId, workflowDefinitionId, workflowDefinitionVersion, context, state } = workflow;
 
   return {
     id,
@@ -135,6 +136,7 @@ export const generateBusiness = ({
         workflowDefinitionId,
         createdAt: faker.date.recent(2),
         projectId: projectId,
+        state: state,
         tags: [StateTag.MANUAL_REVIEW],
       },
     },
@@ -166,10 +168,11 @@ export const generateEndUser = ({
     workflowDefinitionVersion: number;
     context: Prisma.InputJsonValue;
     parentRuntimeId?: string;
+    state: string;
   };
   projectId: string;
 }): Prisma.EndUserCreateInput => {
-  const { workflowDefinitionId, workflowDefinitionVersion, context } = workflow;
+  const { workflowDefinitionId, workflowDefinitionVersion, context, state } = workflow;
   let res: Prisma.EndUserCreateInput = {
     id,
     correlationId,
@@ -192,6 +195,7 @@ export const generateEndUser = ({
           createdAt: faker.date.recent(2),
           projectId: projectId,
           parentRuntimeDataId: workflow.parentRuntimeId,
+          state: state,
           tags: [StateTag.MANUAL_REVIEW],
         },
       },