+++ examples/kyb-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # kyb-app
 
+## 0.1.3
+
+### Patch Changes
+
+- Bump minor version
+- Updated dependencies
+  - @ballerine/ui@0.3.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Version minor bump
+- Updated dependencies
+  - @ballerine/ui@0.3.2
+
 ## 0.1.1
 
 ### Patch Changes